Revert "merged with upstream"
This reverts commit 0c96e9d8391e132476ba99605f9d5cc01791e8e8, reversing
changes made to 72bfd5d1fb149af9703969cd69d6f029d5d24468.

+++ avida-core/source/cpu/cHardwareBase.cc
@@ -243,8 +243,7 @@ bool cHardwareBase::Divide_CheckViable(cAvidaContext& ctx, const int parent_size
   
   // Determine the fitness of this organism as compared to its parent...
   if (m_world->GetTestSterilize() && !phenotype.IsInjected()) {
-    //LZ
-    const double merit_base = phenotype.CalcSizeMerit();
+    const int merit_base = phenotype.CalcSizeMerit();
     const double cur_fitness = merit_base * phenotype.GetCurBonus() / phenotype.GetTimeUsed();
     const double fitness_ratio = cur_fitness / phenotype.GetLastFitness();
     const Apto::Array<int>& childtasks = phenotype.GetCurTaskCount();
